Split assembler and cores modules
